Fix compiler warnings

+++ src/archive.cpp
@@ -51,7 +51,7 @@ namespace librealsense
         const auto texcoords = get_texture_coordinates();
         std::vector<float3> new_vertices;
         std::vector<std::tuple<uint8_t, uint8_t, uint8_t>> new_tex;
-        std::map<int, int> index2reducedIndex;
+        std::map<size_t, size_t> index2reducedIndex;
 
         new_vertices.reserve(get_vertex_count());
         new_tex.reserve(get_vertex_count());
@@ -72,7 +72,7 @@ namespace librealsense
         const auto threshold = 0.05f;
         auto width = video_stream_profile->get_width();
         std::vector<std::tuple<int, int, int>> faces;
-        for (int x = 0; x < width - 1; ++x) {
+        for (size_t x = 0; x < width - 1; ++x) {
             for (int y = 0; y < video_stream_profile->get_height() - 1; ++y) {
                 auto a = y * width + x, b = y * width + x + 1, c = (y + 1)*width + x, d = (y + 1)*width + x + 1;
                 if (vertices[a].z && vertices[b].z && vertices[c].z && vertices[d].z
